Validate and parse sqldialect
